clar: emit clar_suite.h
Old versions of gcc expect us to declare our test functions.  Produce
a clar_suite.h with this information and use it on old versions.

(Do not use it on *new* versions because its timestamp will be
updated whenever a new function is added, causing a complete
recompilation of the test framework.)