diff: provide a flag for whether size is known
Much like we may produce diffs that do not have valid object IDs because
we don't know them (yet), we may also produce diffs that do not have
valid sizes.  Previously, we used a `0` size to indicate unknown, but
this is obviously deficient since we could _know_ that a file is 0
bytes.

Create a new flag that indicates that we positively know the size of the
file.