clone: introduce `no_checkout` option
Using `GIT_CHECKOUT_NONE` as an indication that we shouldn't checkout
during clone saves options(!) but is a very leaky abstraction.
`git_checkout` shouldn't have an idea of being invoked just to do nothing
(why would somebody invoke it then?); instead we should teach clone not
to do the checkout.