core: allow users to configure home directory
Some callers -- like our test suite and the test suites of our language
bindings -- want to isolate the home directory to avoid accidentally
including the executing user's actual home directory data.

Previously, we combined the notion of a home directory and global
configuration -- now that this is separated, we provide users the
ability to configure both.