repo: don't trim trailing slash on root paths
We want to display the safe path (the problematic root path) without a
trailing slash, primarily because that's the path that would need to be
allowlisted. However, if the current repository path is a root path, do not
trim the trailing slash. This ensures that we display '/' on POSIX and a
"proper" drive letter path (eg, 'C:/') on Windows.